Vertex Color Layer Merge
Added the ability to merge layers into a Vertex Color layer.  This basically just merges Height, Thickness, Branch to Frond and Branch Group into a Vertex Color layer and is not customizable yet.

Additionally, added randomization to the branch group generation to make it more natural to use as an offset.